+++ tests/kernel/interrupt/prj.conf
@@ -1,4 +1,3 @@
 CONFIG_ZTEST=y
-CONFIG_IRQ_OFFLOAD=y
 CONFIG_DYNAMIC_INTERRUPTS=y
 CONFIG_MP_NUM_CPUS=1
