+++ tests/kernel/interrupt/src/nested_irq.c
@@ -3,11 +3,12 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "interrupt.h"
+
+#include <ztest.h>
+#include "interrupt_util.h"
 
 #define DURATION 5
 struct k_timer timer;
-struct k_timer irqlock_timer;
 
 /* This tests uses two IRQ lines, selected within the range of IRQ lines
  * available on the target SOC the test executes on (and starting from
@@ -34,11 +35,8 @@ u32_t irq_line_1;
 #define ISR1_PRIO 0
 #endif /* CONFIG_CPU_CORTEX_M */
 
-#define MS_TO_US(ms)  (K_MSEC(ms) * USEC_PER_MSEC)
 volatile u32_t new_val;
 u32_t old_val = 0xDEAD;
-volatile u32_t check_lock_new;
-u32_t check_lock_old = 0xBEEF;
 
 void isr1(void *param)
 {
@@ -130,41 +128,3 @@ void test_nested_isr(void)
 	ztest_test_skip();
 }
 #endif /* NO_TRIGGER_FROM_SW */
-
-static void timer_handler(struct k_timer *timer)
-{
-	ARG_UNUSED(timer);
-	check_lock_new = 0xBEEF;
-	printk("timer fired\n");
-}
-
-void test_prevent_interruption(void)
-{
-	int key;
-
-	printk("locking interrupts\n");
-	key = irq_lock();
-
-	check_lock_new = 0;
-
-	k_timer_init(&irqlock_timer, timer_handler, NULL);
-
-	/* Start the timer and busy-wait for a bit with IRQs locked. The
-	 * timer ought to have fired during this time if interrupts weren't
-	 * locked -- but since they are, check_lock_new isn't updated.
-	 */
-	k_timer_start(&irqlock_timer, DURATION, K_NO_WAIT);
-	k_busy_wait(MS_TO_US(1000));
-	zassert_not_equal(check_lock_new, check_lock_old,
-		"Interrupt locking didn't work properly");
-
-	printk("unlocking interrupts\n");
-	irq_unlock(key);
-
-	k_busy_wait(MS_TO_US(1000));
-
-	zassert_equal(check_lock_new, check_lock_old,
-		"timer should have fired");
-
-	k_timer_stop(&irqlock_timer);
-}
